# src/pkg/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The IPEHR Gateway is an openEHR compliant EHR server implementation that stores encrypted medical data in a Filecoin distributed file storage.",
        "title": "IPEHR Gateway API",
        "contact": {
            "name": "API Support",
            "url": "https://bsn.si/blockchain",
            "email": "support@bsn.si"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.2"
    },
    "host": "gateway.ipehr.org",
    "basePath": "/v1",
    "paths": {
...
        "/user/group": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USER_GROUP"
                ],
                "summary": "Get a list of user groups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.UserGroup"
                            }
                        }
                    },
                    "400": {
                        "description": "The request could not be understood by the server due to incorrect syntax."
                    },
                    "404": {
                        "description": "Is returned when groupID does not exist"
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "USER_GROUP"
                ],
                "summary": "User group create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "User group",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserGroup"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Indicates that the request has succeeded and transaction about create new user group has been created",
                        "schema": {
                            "$ref": "#/definitions/model.UserGroup"
                        },
                        "headers": {
                            "RequestID": {
                                "type": "string",
                                "description": "Request identifier"
                            }
                        }
                    },
                    "400": {
                        "description": "The request could not be understood by the server due to incorrect syntax. The client SHOULD NOT repeat the request without modifications."
                    },
                    "404": {
                        "description": "User with ID not exist"
                    },
                    "409": {
                        "description": "Group with that Name already exist"
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
        "/user/group/{group_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USER_GROUP"
                ],
                "summary": "Get user group by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User group identifier. Example: 7d44b88c-4199-4bad-97dc-d78268e01398",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserGroup"
                        }
                    },
                    "400": {
                        "description": "The request could not be understood by the server due to incorrect syntax."
                    },
                    "403": {
                        "description": "Is returned when userID does not have access to requested group"
                    },
                    "404": {
                        "description": "Is returned when groupID does not exist"
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
        "/user/group/{group_id}/user_add/{user_id}/{access_level}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "USER_GROUP"
                ],
                "summary": "Adding a user to a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the user group",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the user to be added",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Access Level. One of `admin` or `read`",
                        "name": "access_level",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {
                            "RequestID": {
                                "type": "string",
                                "description": "Request identifier"
                            }
                        }
                    },
                    "400": {
                        "description": "The request could not be understood by the server due to incorrect syntax."
                    },
                    "403": {
                        "description": "Authentication required or user does not have access to change the group"
                    },
                    "404": {
                        "description": "Group or adding user is not exist"
                    },
                    "409": {
                        "description": "The user is already a member of a group"
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
        "/user/group/{group_id}/user_remove/{user_id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "USER_GROUP"
                ],
                "summary": "Removing a user from a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the system, typically a reverse domain identifier",
                        "name": "EhrSystemId",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the user group",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The identifier of the user to be removed",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {
                            "RequestID": {
                                "type": "string",
                                "description": "Request identifier"
                            }
                        }
                    },
                    "400": {
                        "description": "The request could not be understood by the server due to incorrect syntax."
                    },
                    "403": {
                        "description": "Authentication required or user does not have access to change the group"
                    },
                    "404": {
                        "description": "Group or adding user is not exist or `user_id` is not the member of the group"
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.UserGroup": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "groupID": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}